+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/features/express/ExpressPaymentPresenter.java
@@ -251,11 +251,12 @@ import java.util.Set;
 
     @Override
     public void cancel() {
-        trackAbort();
+        tracker.trackBack();
         getView().cancel();
     }
 
-    public void trackAbort() {
+    @Override
+    public void onBack() {
         tracker.trackAbort();
     }
 
@@ -449,7 +450,7 @@ import java.util.Set;
                 new FromModalToGenericDialogItem(actionTypeWrapper.getActionType(), behaviour.getModal()).map(modal));
             return true;
         } else if (isMethodSuspended && TextUtil.isNotEmpty(target)) {
-            tracker.trackEvent(new TargetBehaviourEvent(new TargetBehaviourTrackData(behaviourType, target)));
+            new TargetBehaviourEvent(viewTracker, new TargetBehaviourTrackData(behaviourType, target)).track();
             getView().startDeepLink(target);
             return true;
         } else if (isMethodSuspended) {